+++ docs/CHANGELOG.md
@@ -1,3 +1,22 @@
+## v1.5.3 - 2026-05-29
+
+## Updates
+
+- Updated README.md
+- Updated SPEC.md
+- With this update, jeomlang is now truly complete.
+
+---
+
+## v1.5.2 - 2026-05-27
+
+## Updates
+
+- Deleted unnecessary files of vscode extension.
+- Updated grammar.md
+
+---
+
 ## v1.5.1 - 2026-05-25
 
 ## Updates